\name{get.exp}
\alias{get.exp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Extract results from the BayesPrism object
}
\description{
Extract cell type fraction and cell type-specific gene expression estimates from the BayesPrism object
}
\usage{
get.fraction (bp,
              which.theta=c("first", "final"),
              state.or.type=c("state", "type"))

get.exp (bp,
         state.or.type=c("state", "type"),
         cell.name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bp}{A BayesPrism object generated by the run.prism function.}
  \item{which.theta}{A character variable to denote whether to extract restuls from first or final Gibbs sampling.}
  \item{state.or.type}{A character variable to extract restuls from cell type or cell state. We caution the interpretation of cell states information when their transcription are highly co-linear.}
  \item{cell.name}{A character variable or vector to denote the cell type/state to be returned.}
}

\value{
get.fraction returns sample-by-cell type/state matrix

get.exp returns a sample-by-gene-by-cell type/state tensor, if length(cell.name)>1, or a sample-by-gene matrix, if length(cell.name)==1. 
}

