\name{plot.cor.phi}
\alias{plot.cor.phi}
\title{
Plot correlation matrix for the reference profile \eqn{\phi}
}
\description{
This function sum up reads across each cell type/state and drive a gene expression profile matrix (GEP). A pearson correlation matrix between each cell type/state is computed on the log2 transformed GEP. The heatmap generated by this function serves as QC for the cell type/state labels.
}
\usage{
plot.cor.phi (input, 
              input.labels,
              min.exp=3,
              pseudo.min=1E-8, 
              my_palette= colorRampPalette(c("blue", "white", "red"))(n = 254), 
              title="",
              symkey=F,
              symbreaks=F,
              pdf.prefix=NULL,
              ...)
}
\arguments{
  \item{input}{A cell-by-gene count matrix. rownames are cell IDs, while colnames are gene IDs/names. Or a cell type/state-by-gene GEP (non-log scale). rownames are cell state/type names, while colnames are gene IDs/names.}
  \item{input.labels}{A character vector to denote cell stat/type of each cell (if input is a count.matrix) or each row of GEP (if input is a "GEP")}
  \item{min.exp}{Genes expressed in at least this number of cells will be used. Default=3.}
  \item{pseudo.min}{The desired min values to replace zero after normalization. Default=1E-8.}
  \item{my_palette}{The palette for heatmap. Default is blue-white-red.}
  \item{title}{Title of heatmap. Default is none.}
  \item{symkey}{The symkey parameter of heatmap.2. Default=FALSE.}
  \item{symbreaks}{The symbreaks parameter of heatmap.2. Default=FALSE.}
  \item{pdf.prefix}{A character for the prefix of the name of pdf outputs. Default=NULL (plot directly).}
  \item{...}{Other arguments passed to heatmap.2.}
}

